requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
jinja2
aiofiles
python-multipart
email-validator
pydantic-settings
itsdangerous
authlib
python-jose
Babel